+++ src/maggma/builders/__init__.py
@@ -1,0 +1,2 @@
+from maggma.builders.map_builder import MapBuilder, CopyBuilder
+from maggma.builders.group_builder import GroupBuilder
